### 2017-11-04 #450 @whedon
[PRE REVIEW]: Retriever: Data Retrieval Tool
**Submitting author:** @henrykironde (<a href="http://orcid.org/0000-0001-7105-5808">henry senyondo</a>)
**Repository:** <a href="https://github.com/weecology/retriever" target ="_blank">https://github.com/weecology/retriever</a>
**Version:** v2.1.0
