### 2017-11-07 2b0b06f
Download the file to script write path
