+++ retriever/lib/repository.py
@@ -46,7 +46,7 @@ def check_for_updates(quiet=False):
         pkg = os.path.join(os.path.abspath(SCRIPT_WRITE_PATH),
                            "datapackages.yml")
         pkg_url = "https://raw.githubusercontent.com/weecology/retriever/master/scripts/datapackages.yml"
-        urllib.request.urlretrieve(pkg_url, "datapackages.yml")
+        urllib.request.urlretrieve(pkg_url, pkg)
         if not quiet:
             print("Downloading scripts...")
             _update_progressbar(0.0 / float(total_script_count))
