### 2017-11-07 #450 @leeper
@amoeba @henrykironde Review is setup in #451, please take a look at the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) as needed and let me know if you have any questions.
