### 2017-11-13 #451 @amoeba
Grade: 2) Minor revisions

I've had a chance to sit down and found this submission meets or exceeds the submission requirements on essentially every item on the reviewer checklist. It's clear to me what the software should do, it was easy for me to get the software installed and make use of it. On top of that, the documentation is abundant and well-written and it's clear a great deal of care has been put into many aspects of the project. I've checked all but one item off the list and am left with two outstanding issues:

1. The `paper.md` does not appear to follow the BibTeX convention for in-text citations using the BibTeX `@` identifiers and instead just hard-codes the reference. If my assessment is correct, I'd like to see a quick fix made for those.
2. I'm not 100% confident this submission qualifies as research software. It's fundamental tooling for research so I would happily consider it research software. A quick 👍 from an editor or JOSS would suffice.

@leeper can you please take a look at my two remaining issues and let me know what you think?
