### 2017-11-13 afbff12
Use BibTeX @ identifiers
