### 2017-11-13 #451 @henrykironde
Updated to BibTeX citations
