### 2017-11-13 #451 @amoeba
Great, thanks @henrykironde. Relevant checklist item is now checked.

Grade: 1) Accept

Last outstanding issue is just a quick 👍 from @leeper on fit. I have no outstanding issues.
