### 2017-11-14 deef51b
added singleton class for scripts to prevent multiple module loads
