### 2017-11-14 #451 @ethanwhite
Regarding fit, this seems very much in line with other software published in JOSS. E.g.,

* http://joss.theoj.org/papers/421837399efdbef2a248d0cf4a6c1d15
* http://joss.theoj.org/papers/0f59fb7eaeb2004ea510d38c00051dd3

My group is a research lab and this software was built for our research and funded as research software by both NSF and the Moore Foundation.
