### 2017-11-14 #451 @amoeba
I totally agree, @ethanwhite, and imagine @leeper won't disagree either. Thanks for chiming in with those examples. Perhaps the fact that this submission is in the `review` phase is adequate confirmation of fit.
