### 2017-11-15 #451 @leeper
@amoeba Excellent. Thank you for your quick review! Very much appreciated.

@henrykironde This looks ready to me. Great submission and thanks for your quick response to the review.

@arfon Over to you.
