### 2017-11-16 #451 @Deborah-Digges
@arfon I have been affiliated with the University of Florida. Could this affiliation be removed?
