### 2017-11-16 #451 @arfon
@henrykironde - please update the affiliations in `paper.md` in https://github.com/weecology/retriever and let me know when you're ready to have the paper re-generated.
