### 2017-11-17 #451 @arfon
> @arfon I have updated the paper.md

OK thanks. I've re-generated the paper now too: https://github.com/openjournals/joss-papers/blob/8a8bb9cfd7d0d31ba84475e7708df7b4594413af/joss.00451/10.21105.joss.00451.pdf
