### 2017-12-11 #451 @ethanwhite
@arfon - it looks like maybe the regeneration didn't take care of something somewhere because I just noticed that the citation on the website is for the old (incorrect) name.

![screenshot from 2017-12-11 12-40-18](https://user-images.githubusercontent.com/744427/33845154-79688204-de70-11e7-9f56-25f7cb36547f.png)
